+++ src/main/java/com/xzavier0722/mc/plugin/slimefun4/storage/controller/BlockDataConfigWrapper.java
@@ -14,6 +14,7 @@ import org.jetbrains.annotations.ApiStatus;
 @ApiStatus.Internal
 @SlimefunInternal
 @Deprecated
+@SuppressWarnings("deprecation")
 public class BlockDataConfigWrapper extends Config {
     private final ASlimefunDataContainer blockData;