+++ src/main/java/me/mrCookieSlime/Slimefun/api/BlockStorage.java
@@ -23,6 +23,7 @@ import org.bukkit.inventory.ItemStack;
  */
 @Deprecated
 @SlimefunAPI
+@SuppressWarnings("deprecation")
 public class BlockStorage {
 
     private static final Config emptyData = new EmptyBlockData();